requirements.txt
asgiref==3.2.3
dj-database-url==0.5.0
Django==3.0.2
django-ckeditor==5.8.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-six==1.0.4
Pillow==7.0.0
pkg-resources==0.0.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
pytz==2019.3
sqlparse==0.3.0
whitenoise==5.0.1
